portals/requirements_portals.txt
# Portals Affiliate Listener Dependencies
# Install with: pip install -r requirements_portals.txt

# Web3 library for Ethereum interaction
web3>=6.0.0

# Ethereum ABI decoding
eth-abi>=4.0.0

# Environment variable loading
python-dotenv>=1.0.0

# Additional optional dependencies for enhanced functionality
# requests>=2.28.0  # For potential API calls
# pandas>=1.5.0    # For CSV data analysis

